opensles: make the volume property actually work

+++ sys/opensles/openslessink.h
@@ -40,7 +40,7 @@ struct _GstOpenSLESSink
 {
   GstBaseAudioSink sink;
 
-  double volume;
+  gfloat volume;
 };
 
 struct _GstOpenSLESSinkClass
